correciones y reordenamiento ETL
